requirements.txt
requests
setuptools
python-dotenv
customtkinter